+++ tests/Integration/DSC_SqlScript.config.ps1
@@ -28,6 +28,7 @@ else
 
                 Database1Name     = 'ScriptDatabase1'
                 Database2Name     = 'ScriptDatabase2'
+                Database3Name     = '$(DatabaseName)'
 
                 GetSqlScriptPath  = Join-Path -Path $env:SystemDrive -ChildPath ([System.IO.Path]::GetRandomFileName())
                 SetSqlScriptPath  = Join-Path -Path $env:SystemDrive -ChildPath ([System.IO.Path]::GetRandomFileName())
@@ -254,3 +255,65 @@ Configuration DSC_SqlScript_RunSqlScriptAsSqlUser_Config
         }
     }
 }
+
+<#
+    .SYNOPSIS
+        Runs the SQL script with variables disabled.
+
+    .NOTES
+        When the test is run with the DisableVariables parameter the ConfigData variable as it apears in the query will be used as the database name.
+        For example if the ConfigData database name is written as $(DatabaseName) in the T-SQL then the name of the database created will be $(DatabaseName).
+        It might appear odd but it is a valid name.
+#>
+Configuration DSC_SqlScript_RunSqlScriptWithVariablesDisabled_Config
+{
+    Import-DscResource -ModuleName 'SqlServerDsc'
+
+    node $AllNodes.NodeName
+    {
+        SqlScript 'Integration_Test'
+        {
+            ServerName     = $Node.ServerName
+            InstanceName   = $Node.InstanceName
+
+            GetFilePath       = $Node.GetSqlScriptPath
+            TestFilePath      = $Node.TestSqlScriptPath
+            SetFilePath       = $Node.SetSqlScriptPath
+            DisableVariables  = $true
+            QueryTimeout      = 30
+            Credential        = New-Object `
+                -TypeName System.Management.Automation.PSCredential `
+                -ArgumentList @($Node.SqlLogin_UserName, (ConvertTo-SecureString -String $Node.SqlLogin_Password -AsPlainText -Force))
+        }
+    }
+}
+
+<#
+    .SYNOPSIS
+        Remove the database created from the DisabledVariables test.
+
+    .NOTES
+        The current integration test run order has SqlScriptQuery running after SqlScript. Since both the SqlScript and SqlScriptQuery
+        resources use the same database name variable in their ConfigData queries it is imperative this database is absent when
+        the SqlScriptQuery resource runs its DisableVariables integration test.
+#>
+Configuration DSC_SqlScript_RemoveDatabase3_Config
+{
+    Import-DscResource -ModuleName 'SqlServerDsc'
+
+    node $AllNodes.NodeName
+    {
+        SqlDatabase 'RemoveDatabase3'
+        {
+            Ensure               = 'Absent'
+            ServerName           = $Node.ServerName
+            InstanceName         = $Node.InstanceName
+            Name                 = $Node.Database3Name
+
+            PsDscRunAsCredential = New-Object `
+                -TypeName System.Management.Automation.PSCredential `
+                -ArgumentList @($Node.Admin_UserName, (ConvertTo-SecureString -String $Node.Admin_Password -AsPlainText -Force))
+        }
+    }
+}
+