+++ CHANGELOG.md
@@ -55,6 +55,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     resource README.md, and some was added to the schema MOF parameter
     descriptions ([issue #1580](https://github.com/dsccommunity/SqlServerDsc/issues/1580)).
 
+### Added
+
+- SqlScript
+  - Added the DisableVariables parameter ([issue #1422](https://github.com/dsccommunity/SqlServerDsc/issues/1422)).
+- SqlScriptQuery
+  - Added the DisableVariables parameter ([issue #1422](https://github.com/dsccommunity/SqlServerDsc/issues/1422)).
+
 ## [14.0.0] - 2020-06-12
 
 ### Remove